requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-appconf==1.0.5
django-ckeditor==6.6.1
django-countries==7.5.1
django-ipware==5.0.0
django-js-asset==2.0.0
django-modeltranslation==0.18.11
django-recaptcha==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
frozenlist==1.3.3
geoip2==4.7.0
idna==3.4
maxminddb==2.4.0
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyuca==1.2
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uuid==1.30
yarl==1.9.2
